move FileUtils helper class
