+++ Ginger/GingerCoreNET/BPMN/Exportation/BusinessFlowToBPMNExporter.cs
@@ -20,9 +20,9 @@ using Amdocs.Ginger.CoreNET.BPMN.Conversion;
 using Amdocs.Ginger.CoreNET.BPMN.Exceptions;
 using Amdocs.Ginger.CoreNET.BPMN.Models;
 using Amdocs.Ginger.CoreNET.BPMN.Serialization;
+using Amdocs.Ginger.CoreNET.Utility;
 using GingerCore;
 using GingerCore.Activities;
-using GingerUtils;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
